requirements.txt
# Yahoo Fantasy Dashboard - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Dashboard
streamlit>=1.31.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP requests and OAuth
requests>=2.31.0
httpx>=0.26.0
authlib>=1.3.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualization (Plotly instead of matplotlib)
plotly>=5.18.0

# Background Jobs
apscheduler>=3.10.0

# Environment variables
python-dotenv>=1.0.0

# Session management
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.0

# Date/time handling
python-dateutil>=2.8.0
